분리 완료
